roachtest: move load group tests to own file
Release note: None
