+++ pkg/cmd/roachtest/z_cluster_test.go
@@ -349,78 +349,6 @@ func TestClusterMachineType(t *testing.T) {
 	}
 }
 
-func TestLoadGroups(t *testing.T) {
-	cfg := &logger.Config{Stdout: os.Stdout, Stderr: os.Stderr}
-	logger, err := cfg.NewLogger("" /* path */)
-	if err != nil {
-		t.Fatal(err)
-	}
-	for _, tc := range []struct {
-		numZones, numRoachNodes, numLoadNodes int
-		loadGroups                            loadGroupList
-	}{
-		{
-			3, 9, 3,
-			loadGroupList{
-				{
-					option.NodeListOption{1, 2, 3},
-					option.NodeListOption{4},
-				},
-				{
-					option.NodeListOption{5, 6, 7},
-					option.NodeListOption{8},
-				},
-				{
-					option.NodeListOption{9, 10, 11},
-					option.NodeListOption{12},
-				},
-			},
-		},
-		{
-			3, 9, 1,
-			loadGroupList{
-				{
-					option.NodeListOption{1, 2, 3, 4, 5, 6, 7, 8, 9},
-					option.NodeListOption{10},
-				},
-			},
-		},
-		{
-			4, 8, 2,
-			loadGroupList{
-				{
-					option.NodeListOption{1, 2, 3, 4},
-					option.NodeListOption{9},
-				},
-				{
-					option.NodeListOption{5, 6, 7, 8},
-					option.NodeListOption{10},
-				},
-			},
-		},
-	} {
-		t.Run(fmt.Sprintf("%d/%d/%d", tc.numZones, tc.numRoachNodes, tc.numLoadNodes),
-			func(t *testing.T) {
-				c := &clusterImpl{t: testWrapper{T: t}, l: logger, spec: spec.MakeClusterSpec(spec.GCE, "", tc.numRoachNodes+tc.numLoadNodes)}
-				lg := makeLoadGroups(c, tc.numZones, tc.numRoachNodes, tc.numLoadNodes)
-				require.EqualValues(t, lg, tc.loadGroups)
-			})
-	}
-	t.Run("panics with too many load nodes", func(t *testing.T) {
-		require.Panics(t, func() {
-
-			numZones, numRoachNodes, numLoadNodes := 2, 4, 3
-			makeLoadGroups(nil, numZones, numRoachNodes, numLoadNodes)
-		}, "Failed to panic when number of load nodes exceeded number of zones")
-	})
-	t.Run("panics with unequal zones per load node", func(t *testing.T) {
-		require.Panics(t, func() {
-			numZones, numRoachNodes, numLoadNodes := 4, 4, 3
-			makeLoadGroups(nil, numZones, numRoachNodes, numLoadNodes)
-		}, "Failed to panic when number of zones is not divisible by number of load nodes")
-	})
-}
-
 func TestCmdLogFileName(t *testing.T) {
 	ts := time.Date(2000, 1, 1, 15, 4, 12, 0, time.Local)
